v3

+++ ControlPanelDynamic/src/main/java/com/sbt/test/entities/EventType.java
@@ -21,6 +21,6 @@ public class EventType {
 
     private String name;
 
-    long operatingRatio;
+    double operatingRatio;
 
 }
